+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **38_factos_costs** sticky_labor realization: included option to set a labor share target
 
 ### added
--
+- **scripts** added start script which starts an empty model just regenerating a previous run
 
 ### removed
 -